# apps/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/inventory/{restaurantId}/summary": {
      "get": {
        "operationId": "InventoryController_getInventorySummary",
        "summary": "Get inventory summary statistics",
        "parameters": [
          {
            "name": "restaurantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventorySummaryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "inventory"
        ]
      }
    },
...
    "/api/v1/inventory-ledger/transactions": {
      "post": {
        "operationId": "InventoryLedgerController_createTransaction",
        "summary": "Record a new inventory transaction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryTransactionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "inventory-ledger"
        ]
      },
      "get": {
        "operationId": "InventoryLedgerController_listTransactions",
        "summary": "List inventory transactions with filters",
        "parameters": [
          {
            "name": "inventoryId",
            "required": false,
            "in": "query",
            "description": "Filter by inventory item ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "wineId",
            "required": false,
            "in": "query",
            "description": "Filter by wine ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transactionType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "sale",
                "purchase",
                "adjustment",
                "transfer",
                "waste",
                "return",
                "comp",
                "reconciliation",
                "initial",
                "correction"
              ],
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pos",
                "manual",
                "order",
                "mobile_count",
                "reconciliation",
                "system",
                "import",
                "api"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date (ISO)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date (ISO)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns transactions list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "inventory-ledger"
        ]
      }
    },
    "/api/v1/inventory-ledger/transactions/bulk": {
      "post": {
        "operationId": "InventoryLedgerController_createBulkTransactions",
        "summary": "Record multiple inventory transactions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transactions recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkTransactionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "inventory-ledger"
        ]
      }
    },
    "/api/v1/inventory-ledger/transactions/{transactionId}": {
      "get": {
        "operationId": "InventoryLedgerController_getTransaction",
        "summary": "Get a specific transaction",
        "parameters": [
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "description": "Transaction ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryTransactionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "inventory-ledger"
        ]
      }
    },
    "/api/v1/inventory-ledger/inventory/{inventoryId}/balance": {
      "get": {
        "operationId": "InventoryLedgerController_getBalanceAt",
        "summary": "Get inventory balance at a point in time",
        "parameters": [
          {
            "name": "inventoryId",
            "required": true,
            "in": "path",
            "description": "Inventory item ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asOf",
            "required": true,
            "in": "query",
            "description": "Point in time (ISO timestamp)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stockType",
            "required": false,
            "in": "query",
            "schema": {
              "default": "live",
              "enum": [
                "live",
                "shadow",
                "reserved"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryBalanceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "inventory-ledger"
        ]
      }
    },
    "/api/v1/inventory-ledger/inventory/{inventoryId}/history": {
      "get": {
        "operationId": "InventoryLedgerController_getTransactionHistory",
        "summary": "Get transaction history for an inventory item",
        "parameters": [
          {
            "name": "inventoryId",
            "required": true,
            "in": "path",
            "description": "Inventory item ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "Number of days to look back (default: 30)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns transaction history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryTransactionResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "inventory-ledger"
        ]
      }
    },
    "/api/v1/inventory-ledger/summary": {
      "get": {
        "operationId": "InventoryLedgerController_getTransactionSummary",
        "summary": "Get transaction summary for a date range",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "description": "Start date (ISO)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "description": "End date (ISO)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns transaction summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSummaryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "inventory-ledger"
        ]
      }
    },
    "/api/v1/inventory-ledger/inventory/{inventoryId}/reconcile": {
      "post": {
        "operationId": "InventoryLedgerController_reconcileInventory",
        "summary": "Reconcile inventory with physical count",
        "parameters": [
          {
            "name": "inventoryId",
            "required": true,
            "in": "path",
            "description": "Inventory item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Reconciliation recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryTransactionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "inventory-ledger"
        ]
      }
    },
...
  },
  "info": {
    "title": "WineOps AI API",
    "description": "\n## Overview\n\nReal-time API for Restaurant Wine Inventory & Procurement Management.\n\n### Features\n\n- **Dashboard Aggregation**: Single endpoint for all dashboard data (API Bus pattern)\n- **One-Tap Actions**: Quick approval workflows for managers\n- **Toast POS Integration**: Proxy endpoints for Toast API\n- **Real-time Updates**: WebSocket support for live data sync\n- **Rate Limiting**: Per-user and per-restaurant quotas\n\n### Authentication\n\nAll endpoints (except public ones) require a Bearer token in the Authorization header.\n\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n### Rate Limits\n\n| Endpoint Type | Limit | Window |\n|--------------|-------|--------|\n| Default | 100 | 60s |\n| Auth | 10 | 60s |\n| Upload | 10 | 300s |\n| AI | 20 | 60s |\n\n### WebSocket Events\n\nConnect to `ws://localhost:4000` and join rooms:\n- `restaurant:{restaurantId}` - Restaurant-wide events\n- `user:{userId}` - User-specific events\n\nEvents:\n- `notification` - New notifications\n- `inventory_update` - Inventory changes\n- `order_update` - Order status changes\n- `one_tap_action` - One-tap action updates\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "WineOps Team",
      "url": "https://wineops.ai",
      "email": "support@wineops.ai"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://wineops.ai/license"
    }
  },
...
  "servers": [
    {
...
    },
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "InventorySummaryResponseDto": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "number",
            "description": "Total number of inventory items"
          },
          "totalBottles": {
            "type": "number",
            "description": "Total bottles across all items"
          },
          "lowStockCount": {
            "type": "number",
            "description": "Number of low stock items"
          },
          "criticalCount": {
            "type": "number",
            "description": "Number of out of stock items"
          },
          "healthyCount": {
            "type": "number",
            "description": "Number of healthy stock items"
          },
          "toastMappedCount": {
            "type": "number",
            "description": "Number of items mapped to Toast"
          },
          "toastUnmappedCount": {
            "type": "number",
            "description": "Number of items not mapped to Toast"
          }
        },
        "required": [
          "totalItems",
          "totalBottles",
          "lowStockCount",
          "criticalCount",
          "healthyCount"
        ]
      },
...
      "CreateInventoryTransactionDto": {
        "type": "object",
        "properties": {
          "inventoryId": {
            "type": "string",
            "description": "Inventory item ID"
          },
          "wineId": {
            "type": "string",
            "description": "Wine ID"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "sale",
              "purchase",
              "adjustment",
              "transfer",
              "waste",
              "return",
              "comp",
              "reconciliation",
              "initial",
              "correction"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "pos",
              "manual",
              "order",
              "mobile_count",
              "reconciliation",
              "system",
              "import",
              "api"
            ]
          },
          "quantityChange": {
            "type": "number",
            "description": "Quantity change (positive = increase, negative = decrease)"
          },
          "stockType": {
            "type": "string",
            "enum": [
              "live",
              "shadow",
              "reserved"
            ],
            "default": "live"
          },
          "referenceType": {
            "type": "string",
            "description": "Reference type (e.g., order, pos_transaction)"
          },
          "referenceId": {
            "type": "string",
            "description": "Reference ID"
          },
          "posTransactionId": {
            "type": "string",
            "description": "POS transaction ID"
          },
          "orderId": {
            "type": "string",
            "description": "Procurement order ID"
          },
          "fromLocationId": {
            "type": "string",
            "description": "Source location ID (for transfers)"
          },
          "toLocationId": {
            "type": "string",
            "description": "Destination location ID (for transfers)"
          },
          "unitCost": {
            "type": "number",
            "description": "Unit cost at time of transaction"
          },
          "reason": {
            "type": "string",
            "description": "Reason for the transaction"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata"
          }
        },
        "required": [
          "inventoryId",
          "wineId",
          "transactionType",
          "source",
          "quantityChange"
        ]
      },
      "InventoryTransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "restaurantId": {
            "type": "string"
          },
          "inventoryId": {
            "type": "string"
          },
          "wineId": {
            "type": "string"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "sale",
              "purchase",
              "adjustment",
              "transfer",
              "waste",
              "return",
              "comp",
              "reconciliation",
              "initial",
              "correction"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "pos",
              "manual",
              "order",
              "mobile_count",
              "reconciliation",
              "system",
              "import",
              "api"
            ]
          },
          "quantityChange": {
            "type": "number"
          },
          "quantityBefore": {
            "type": "number"
          },
          "quantityAfter": {
            "type": "number"
          },
          "stockType": {
            "type": "string",
            "enum": [
              "live",
              "shadow",
              "reserved"
            ]
          },
          "referenceType": {
            "type": "string"
          },
          "referenceId": {
            "type": "string"
          },
          "posTransactionId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "fromLocationId": {
            "type": "string"
          },
          "toLocationId": {
            "type": "string"
          },
          "unitCost": {
            "type": "number"
          },
          "totalCost": {
            "type": "number"
          },
          "performedBy": {
            "type": "string"
          },
          "performedByType": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "transactionDate": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "restaurantId",
          "inventoryId",
          "wineId",
          "transactionType",
          "source",
          "quantityChange",
          "quantityBefore",
          "quantityAfter",
          "stockType",
          "performedByType",
          "transactionDate",
          "createdAt"
        ]
      },
      "BulkTransactionDto": {
        "type": "object",
        "properties": {
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateInventoryTransactionDto"
            }
          },
          "correlationId": {
            "type": "string",
            "description": "Correlation ID for all transactions"
          }
        },
        "required": [
          "transactions"
        ]
      },
      "BulkTransactionResponseDto": {
        "type": "object",
        "properties": {
          "successCount": {
            "type": "number"
          },
          "failedCount": {
            "type": "number"
          },
          "createdIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "successCount",
          "failedCount",
          "createdIds",
          "errors"
        ]
      },
      "TransactionsListResponseDto": {
        "type": "object",
        "properties": {
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryTransactionResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "transactions",
          "total",
          "page",
          "limit",
          "hasMore"
        ]
      },
      "InventoryBalanceResponseDto": {
        "type": "object",
        "properties": {
          "inventoryId": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "asOf": {
            "type": "string"
          },
          "stockType": {
            "type": "string",
            "enum": [
              "live",
              "shadow",
              "reserved"
            ]
          }
        },
        "required": [
          "inventoryId",
          "balance",
          "asOf",
          "stockType"
        ]
      },
      "TransactionSummaryResponseDto": {
        "type": "object",
        "properties": {
          "restaurantId": {
            "type": "string"
          },
          "period": {
            "type": "string"
          },
          "totalIn": {
            "type": "number"
          },
          "totalOut": {
            "type": "number"
          },
          "netChange": {
            "type": "number"
          },
          "transactionCount": {
            "type": "number"
          },
          "byType": {
            "type": "object"
          },
          "bySource": {
            "type": "object"
          }
        },
        "required": [
          "restaurantId",
          "period",
          "totalIn",
          "totalOut",
          "netChange",
          "transactionCount",
          "byType",
          "bySource"
        ]
      },
...
    }
  }
}